Finished chapter 5 - ui decorations

+++ core/src/main/groovy/com/packtpub/libgdx/canyonbunny/game/objects/Clouds.groovy
@@ -70,7 +70,7 @@ class Clouds extends AbstractGameObject {
 
         @Override
         public void render(SpriteBatch batch) {
-            batch.draw(region.texture, position.x + origin.x, position.y + origin.y, origin.x, origin.y, dimension.x, dimension.y, scale.x, scale.y, rotation, region.regionX, region.regionY, region.regionWidth, region.regionHeight, false, false)
+            batch.draw(region.texture, (position.x + origin.x) as float, (position.y + origin.y) as float, origin.x, origin.y, dimension.x, dimension.y, scale.x, scale.y, rotation, region.regionX, region.regionY, region.regionWidth, region.regionHeight, false, false)
         }
     }
 }
