+++ core/src/main/groovy/com/packtpub/libgdx/canyonbunny/game/objects/WaterOverlay.groovy
@@ -17,16 +17,16 @@ class WaterOverlay extends AbstractGameObject {
     }
 
     private void init() {
-        dimension.set(length * 10, 3)
+        dimension.set((length * 10) as float, 3f)
 
         regWaterOverlay = Assets.instance.levelDecoration.waterOverlay
 
-        origin.x = -dimension.x / 2
+        origin.x = (-dimension.x / 2) as float
     }
 
     @Override
     public void render(SpriteBatch batch) {
         TextureRegion reg = regWaterOverlay
-        batch.draw(reg.texture, position.x + origin.x, position.y + origin.y, origin.x, origin.y, dimension.x, dimension.y, scale.x, scale.y, rotation, reg.regionX, reg.regionY, reg.regionWidth, reg.regionHeight, false, false)
+        batch.draw(reg.texture, (position.x + origin.x) as float, (position.y + origin.y) as float, origin.x, origin.y, dimension.x, dimension.y, scale.x, scale.y, rotation, reg.regionX, reg.regionY, reg.regionWidth, reg.regionHeight, false, false)
     }
 }
